+++ src/main/java/ulaval/glo2003/controllers/product/ProductResource.java
@@ -60,8 +60,8 @@ public class ProductResource {
     public Response getFilteredProducts(@QueryParam("sellerId") String sellerId,
                                         @QueryParam("title") String title,
                                         @QueryParam("categories") List<String> categories,
-                                        @QueryParam("minPrice") Float minPrice,
-                                        @QueryParam("maxPrice") Float maxPrice) {
+                                        @QueryParam("minPrice") Double minPrice,
+                                        @QueryParam("maxPrice") Double maxPrice) {
         try {
             List<SellerProduct> sellersProducts = productService.getFilteredProducts(sellerId,
                                                                                      title,
