+++ src/test/java/ulaval/glo2003/controllers/seller/dtos/SellerPresenterTest.java
@@ -26,7 +26,7 @@ public class SellerPresenterTest {
     private static final String SELLER_BIO = "testSellerBio";
     private static final String PRODUCT_TITLE = "testProductTitle";
     private static final String PRODUCT_DESCRIPTION = "testProductDescription";
-    private static final Float PRODUCT_SUGGESTED_PRICE = 1.0f;
+    private static final Double PRODUCT_SUGGESTED_PRICE = 1.;
     private static final List<ProductCategory> PRODUCT_CATEGORIES = new ArrayList<>() {
         {
             add(ProductCategory.APPAREL);
